requirements.txt
# Requirements file

tkinter
requests
pillow
sqlite3